fixed admin  authcontrollers

+++ backend/Controllers/categoryController.js
@@ -2,22 +2,28 @@ import { Category } from "../Models/catgoryModel.js"
 
 export const addCategory = async (req, res) => {
     const { categoryName, slug } = req.body;
-    try {
-        const existingCategory = await Category.findOne({ categoryName: categoryName });
-        if (existingCategory) {
-            return res.status(401).json({ message: "Category Already Exists" })
-        } else {
-            const newcategory = new Category({
-                categoryName: categoryName,
-                slug: slug
-            })
-            await newcategory.save()
-            return res.status(201).json({ message: "Category Added Successfully" })
-        }
+    const { role } = req.user;
+    if (role === "admin") {
+        try {
+            const existingCategory = await Category.findOne({ categoryName: categoryName });
+            if (existingCategory) {
+                return res.status(401).json({ message: "Category Already Exists" })
+            } else {
+                const newcategory = new Category({
+                    categoryName: categoryName,
+                    slug: slug
+                })
+                await newcategory.save()
+                return res.status(201).json({ message: "Category Added Successfully" })
+            }
 
-    } catch (error) {
-        res.status(500).json({ message: "Server Error", error })
-        console.log(error)
+        } catch (error) {
+            res.status(500).json({ message: "Server Error", error })
+            console.log(error)
+        }
+    } else {
+        return res.status(401).json({ message: "Unauthorised User" })
+
     }
 }
 
@@ -36,36 +42,48 @@ export const getCategory = async (req, res) => {
 }
 
 export const editCategory = async (req, res) => {
-    try {
-        const { _id, categoryName, slug } = req.body
-        const category = await Category.findOne({ _id: _id });
-        if (category) {
-            category.categoryName = categoryName;
-            category.slug = slug;
-            await category.save()
-            return res.status(201).json({message:"Category Updated Successfully"})
-        }else{
-            return res.status(401).json({message:"Uncaught Error"})
-        }
+    const { role } = req.user;
+    if (role === "admin") {
 
-    } catch (error) {
-        res.status(500).json({ message: "Server Error", error })
+        try {
+            const { _id, categoryName, slug } = req.body
+            const category = await Category.findOne({ _id: _id });
+            if (category) {
+                category.categoryName = categoryName;
+                category.slug = slug;
+                await category.save()
+                return res.status(201).json({ message: "Category Updated Successfully" })
+            } else {
+                return res.status(401).json({ message: "Uncaught Error" })
+            }
+
+        } catch (error) {
+            res.status(500).json({ message: "Server Error", error })
+        }
+    } else {
+        return res.status(401).json({ message: "Unauthorised User" })
     }
 }
 
 export const deleteCategory = async (req, res) => {
-    try {
-        const {categoryId} = req.body
-        if (categoryId) {
-            const category = await Category.findByIdAndDelete({ _id: categoryId });
-            if(category){
-                return res.status(201).json({message:"Category Deleted Successfully"})
+    const { role } = req.user;
+    if (role === "admin") {
+        try {
+            const { categoryId } = req.body
+            if (categoryId) {
+                const category = await Category.findByIdAndDelete({ _id: categoryId });
+                if (category) {
+                    return res.status(201).json({ message: "Category Deleted Successfully" })
+                }
+            } else {
+                return res.status(401).json({ message: "Uncaught Error" })
             }
-        }else{
-            return res.status(401).json({message:"Uncaught Error"})
-        }
 
-    } catch (error) {
-        res.status(500).json({ message: "Server Error", error })
+        } catch (error) {
+            res.status(500).json({ message: "Server Error", error })
+        }
+    } else {
+        return res.status(401).json({ message: "Unauthorised User" })
     }
+
 }
